Teeny bit of work to get it so my unit-test target works.
