+++ sdk/appconfiguration/app-configuration/test/helpers.spec.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { checkAndFormatIfAndIfNoneMatch } from "../src/internal/helpers"
+import { checkAndFormatIfAndIfNoneMatch, formatWildcards, extractAfterTokenFromNextLink, quoteETag } from "../src/internal/helpers"
 import * as assert from "assert";
 
 describe("helper methods", () => {
@@ -35,3 +35,70 @@ describe("helper methods", () => {
     }), /onlyIfChanged and onlyIfUnchanged are mutually-exclusive/);    
   });
 })
+
+
+
+describe("quoteETag", () => {
+  it("undefined", () => {
+    assert.equal(
+      undefined,
+      quoteETag(undefined)
+    );
+
+    assert.equal(
+      '"etagishere"',
+      quoteETag("etagishere")
+    );
+
+    assert.equal(
+      "'etagishere'",
+      quoteETag("'etagishere'")
+    );
+
+    assert.equal(
+      "*",
+      quoteETag("*")
+    );
+  });
+});
+
+describe("formatWildcards", () => {
+  it("undefined", () => {
+    const result = formatWildcards({
+      keys: undefined,
+      labels: undefined
+    });
+
+    assert.ok(!result.key);
+    assert.ok(!result.label);
+  });
+
+  it("single values only", () => {
+    const result = formatWildcards({
+      keys: ["key1"],
+      labels: ["label1"]
+    });
+
+    assert.equal("key1", result.key);
+    assert.equal("label1", result.label);
+  });
+
+  it("multiple values", () => {
+    const result = formatWildcards({
+      keys: ["key1", "key2"],
+      labels: ["label1", "label2"]
+    });
+
+    assert.equal("key1,key2", result.key);
+    assert.equal("label1,label2", result.label);
+  });
+});
+
+describe("extractAfterTokenFromNextLink", () => {
+  it("token is extracted and properly unescaped", () => {
+    let token = extractAfterTokenFromNextLink(
+      "/kv?key=someKey&api-version=1.0&after=bGlah%3D"
+    );
+    assert.equal("bGlah=", token);
+  });
+});
