+++ sdk/appconfiguration/app-configuration/test/samples.spec.ts
@@ -2,8 +2,13 @@
 // Licensed under the MIT license.
 
 import { runAll } from "../samples";
+import { getConnectionStringFromEnvironment } from './testHelpers';
 
 describe("AppConfiguration samples", () => {
+    before(function () {
+        getConnectionStringFromEnvironment() || this.skip();
+    });
+
     it("Make sure all the samples build and run", async () => {
         await runAll();
     });
